[TSP] add tsp solver

+++ src/main/kotlin/ru/sfedu/geo/model/Point.kt
@@ -4,8 +4,8 @@ import jakarta.persistence.Embeddable
 
 @Embeddable
 data class Point(
-    val lat: Float? = null,
-    val long: Float? = null,
+    val lat: Double? = null,
+    val long: Double? = null,
 ) {
     override fun toString(): String {
         return "($lat, $long)"
